refactor: rename blog detail and collection query functions for consistency and update ArticleDetail component to accept type prop
