+++ app/blog/[slug]/page.tsx
@@ -3,7 +3,7 @@ import { ApiPath, apiRequest } from "@/utils/apiClient";
 import { ArticleDetailResponse } from "@/types/responseTypes";
 import HeroSection from "@/components/media/article-detail/HeroSection";
 import ArticleDetail from "@/components/media/article-detail/ArticleDetail";
-import { getArticleDetailQuery } from "@/utils/blogDetailQuery";
+import { getArticleDetailQuery } from "@/utils/articlesDetailQuery";
 
 // Menandai komponen sebagai async component
 const BlogDetail = async ({
@@ -27,7 +27,10 @@ const BlogDetail = async ({
     return (
       <section>
         <HeroSection data={blogDetailData} />
-        <ArticleDetail content={blogDetailData.content} />
+        <ArticleDetail
+          content={blogDetailData.content}
+          type={blogDetailData.type}
+        />
       </section>
     );
   } catch (error) {
